Added a stupid hack to fix startpos
The Player did not appear to start in the correct position. To alivate
this issue, I have made it attempt to move to the correct position 3
times. This is dumb and shouldn't work, but it does. It is inactive
until the 3rd run, then it appears. I am sorry, but we have a deadline
to meet.